fix: create session wrapper
