+++ moqt/src/modules/moqt/sessions.rs
@@ -1,3 +1,4 @@
+pub(crate) mod inner_session;
 pub(crate) mod session;
 pub(crate) mod session_creator;
 pub(crate) mod session_message_resolver;
