+++ moqt/src/modules/moqt/sessions/session_creator.rs
@@ -1,10 +1,10 @@
 use std::net::SocketAddr;
-use std::sync::Arc;
 
 use crate::modules::moqt::control_receiver::ControlReceiver;
 use crate::modules::moqt::control_sender::ControlSender;
 use crate::modules::moqt::enums::ReceiveEvent;
 use crate::modules::moqt::protocol::TransportProtocol;
+use crate::modules::moqt::sessions::inner_session::InnerSession;
 use crate::modules::moqt::sessions::session::Session;
 use crate::modules::transport::transport_connection::TransportConnection;
 use crate::modules::transport::transport_connection_creator::TransportConnectionCreator;
@@ -18,7 +18,7 @@ impl<T: TransportProtocol> SessionCreator<T> {
         &self,
         remote_address: SocketAddr,
         host: &str,
-    ) -> anyhow::Result<Arc<Session<T>>> {
+    ) -> anyhow::Result<Session<T>> {
         let transport_conn = self
             .transport_creator
             .create_new_transport(remote_address, host)
@@ -28,20 +28,22 @@ impl<T: TransportProtocol> SessionCreator<T> {
         let (sender, _) = tokio::sync::broadcast::channel::<ReceiveEvent>(16);
         let moqt_sender = ControlSender::<T> { send_stream };
         let moqt_receiver = ControlReceiver::new::<T>(receive_stream, sender.clone());
-        Session::<T>::client(transport_conn, moqt_sender, moqt_receiver, sender)
+        let inner = InnerSession::<T>::client(transport_conn, moqt_sender, moqt_receiver, sender)
             .await
-            .inspect(|_| tracing::info!("Session has been created."))
+            .inspect(|_| tracing::info!("Session has been created."))?;
+        Ok(Session::<T>::new(inner))
     }
 
-    pub(crate) async fn accept_new_connection(&mut self) -> anyhow::Result<Arc<Session<T>>> {
+    pub(crate) async fn accept_new_connection(&mut self) -> anyhow::Result<Session<T>> {
         let transport_conn = self.transport_creator.accept_new_transport().await?;
         let (send_stream, receive_stream) = transport_conn.accept_bi().await?;
         // 16 means the number of messages can be stored in the channel.
         let (sender, _) = tokio::sync::broadcast::channel::<ReceiveEvent>(16);
         let moqt_sender = ControlSender { send_stream };
         let moqt_receiver = ControlReceiver::new::<T>(receive_stream, sender.clone());
-        Session::<T>::server(transport_conn, moqt_sender, moqt_receiver, sender)
+        let inner = InnerSession::<T>::server(transport_conn, moqt_sender, moqt_receiver, sender)
             .await
-            .inspect(|_| tracing::info!("Session has been created."))
+            .inspect(|_| tracing::info!("Session has been created."))?;
+        Ok(Session::<T>::new(inner))
     }
 }
